feed property wrapper implementation
